sizes of datatypes

+++ 0x00-hello_world/6-size.c
@@ -9,13 +9,13 @@ int main(void)
 	char my_char;
 	float my_float;
 	int my_int;
-	double my_long_int;
+	long my_long_int;
 	long long my_long_long_int;
 
-	printf("Size of a char: %lu bytes(s)\n", sizeof(my_char));
-	printf("Size of a float: %lu bytes(s)\n", sizeof(my_float));
-	printf("Size of an int: %lu bytes(s)\n", sizeof(my_int));
-	printf("Size of a long int: %lu bytes(s)\n", sizeof(my_long_int));
-	printf("Size of a long long int: %lu bytes(s)\n", sizeof(my_long_long_int));
+	printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(my_char));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(my_float));
+	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(my_int));
+	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(my_long_int));
+	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(my_long_long_int));
 	return (0);
 }
